+++ buildpacks/nodejs-function-invoker/src/main.rs
@@ -1,6 +1,6 @@
 use crate::function::{
-    get_declared_runtime_package_version, get_main, is_function, ExplicitRuntimeDependencyError,
-    MainError,
+    check_minumum_node_version, get_declared_runtime_package_version, get_main, is_function,
+    ExplicitRuntimeDependencyError, MainError, MinimumNodeVersionError,
 };
 use crate::layers::{
     RuntimeLayer, RuntimeLayerError, ScriptLayer, ScriptLayerError, NODEJS_RUNTIME_SCRIPT,
@@ -73,6 +73,9 @@ impl Buildpack for NodeJsInvokerBuildpack {
         let metadata_runtime = &context.buildpack_descriptor.metadata.runtime;
         let package_name = &metadata_runtime.package_name;
         let package_version = &metadata_runtime.package_version;
+
+        check_minumum_node_version(&app_dir)
+            .map_err(NodeJsInvokerBuildpackError::MinimumNodeVersion)?;
 
         log_info("Checking for function file");
         get_main(app_dir).map_err(NodeJsInvokerBuildpackError::MainFunction)?;
@@ -151,6 +154,12 @@ impl Buildpack for NodeJsInvokerBuildpack {
                             err_string,
                         );
                     }
+                    NodeJsInvokerBuildpackError::MinimumNodeVersion(_) => {
+                        log_error(
+                            "Node.js Function Invoker minimum Node.js version error",
+                            err_string,
+                        );
+                    }
                 }
             },
             error,
@@ -160,14 +169,16 @@ impl Buildpack for NodeJsInvokerBuildpack {
 
 #[derive(Error, Debug)]
 enum NodeJsInvokerBuildpackError {
-    #[error("{0}")]
+    #[error(transparent)]
     MainFunction(#[from] MainError),
-    #[error("{0}")]
+    #[error(transparent)]
     RuntimeLayer(#[from] RuntimeLayerError),
-    #[error("{0}")]
+    #[error(transparent)]
     ScriptLayer(#[from] ScriptLayerError),
-    #[error("{0}")]
+    #[error(transparent)]
     ExplicitRuntimeDependencyFunction(#[from] ExplicitRuntimeDependencyError),
+    #[error(transparent)]
+    MinimumNodeVersion(#[from] MinimumNodeVersionError),
 }
 
 impl From<NodeJsInvokerBuildpackError> for libcnb::Error<NodeJsInvokerBuildpackError> {