+++ common/nodejs-utils/src/vrs.rs
@@ -104,7 +104,7 @@ impl Requirement {
     }
 
     #[must_use]
-    pub(crate) fn satisfies(&self, ver: &Version) -> bool {
+    pub fn satisfies(&self, ver: &Version) -> bool {
         self.0.satisfies(&ver.0)
     }
 }